Move cipher crate into ternoa_implementation according to new structure. See #16
