+++ client/src/ternoa_implementation/cipher.rs
