+++ client/src/ternoa_implementation/mod.rs
@@ -1,0 +1,4 @@
+pub mod cipher;
+
+#[cfg(test)]
+mod tests;
